Fix unnecessary type casting
Reviewed-on: http://review.couchbase.org/82659
Tested-by: Build Bot <build@couchbase.com>

+++ perfrunner/workloads/ycsb.py
@@ -9,9 +9,11 @@ def ycsb_data_load(workload_settings: PhaseSettings,
     soe_params = None
     if workload_settings.recorded_load_cache_size:
         restart_memcached()
-        soe_params = {'insertstart': int((instance + 1) *
-                                         workload_settings.inserts_per_workerinstance),
-                      'recorded_load_cache_size': workload_settings.recorded_load_cache_size}
+        soe_params = {
+            'insertstart': (instance + 1) * workload_settings.inserts_per_workerinstance,
+            'recorded_load_cache_size': workload_settings.recorded_load_cache_size,
+        }
+
     run_ycsb(host=target.node,
              bucket=target.bucket,
              password=target.password,
@@ -29,9 +31,10 @@ def ycsb_workload(workload_settings: PhaseSettings,
                   instance: int):
     soe_params = None
     if workload_settings.recorded_load_cache_size:
-        soe_params = {'insertstart': int((instance + 1) *
-                                         workload_settings.inserts_per_workerinstance),
-                      'recorded_load_cache_size': workload_settings.recorded_load_cache_size}
+        soe_params = {
+            'insertstart': (instance + 1) * workload_settings.inserts_per_workerinstance,
+            'recorded_load_cache_size': workload_settings.recorded_load_cache_size,
+        }
 
     run_ycsb(host=target.node,
              bucket=target.bucket,
